+++ CellularAutomata/src/cellularautomata/automata/IntNearAether1_3D.java
@@ -417,7 +417,7 @@ public class IntNearAether1_3D implements SymmetricEvolvingIntGrid3D, Serializab
 	 * 
 	 * @return the value at the origin at step 0
 	 */
-	public long getInitialValue() {
+	public int getInitialValue() {
 		return initialValue;
 	}
