+++ CellularAutomata/src/cellularautomata/automata/IntNearAether2_3D.java
@@ -403,7 +403,7 @@ public class IntNearAether2_3D implements SymmetricEvolvingIntGrid3D, Serializab
 	 * 
 	 * @return the value at the origin at step 0
 	 */
-	public long getInitialValue() {
+	public int getInitialValue() {
 		return initialValue;
 	}
